# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestJobView": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "logs"
        ],
        "title": "IngestJobView",
        "type": "object"
      },
      "IngestRequest": {
        "description": "Ingest a local bundle directory into Your Own (builds its fused image).",
        "properties": {
          "bundle_dir": {
            "title": "Bundle Dir",
            "type": "string"
          }
        },
        "required": [
          "bundle_dir"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "IngestStarted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "IngestStarted",
        "type": "object"
      },
...
      "PullJobStarted": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "PullJobStarted",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "XORCISE",
    "version": "0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/missions/ingest": {
      "post": {
        "description": "Start a local-bundle ingest and return a job id immediately.\n\nThe host-local Docker build can take minutes \u2014 far past the CLI's read timeout and long enough\nto hang a browser request \u2014 so it runs on a background task; poll GET /missions/ingest/{id}\nfor status + progress logs (the CLI and the GUI both do). Routes the install through the\ncontroller so the server remains the single writer (parity with /pull).",
        "operationId": "ingest_api_missions_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/ingest/{job_id}": {
      "get": {
        "description": "Ingest job status + progress logs. `since` = log-index cursor (default 0=all).",
        "operationId": "ingest_status_api_missions_ingest__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Since",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Status",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/pull-jobs": {
      "get": {
        "description": "The active ('pulling') job for a mission, or null \u2014 lets a reloaded GUI resume its\nprogress bar and keeps a second view's Pull button disabled.",
        "operationId": "active_pull_job_api_missions_pull_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PullJobView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Active Pull Job Api Missions Pull Jobs Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Pull Job",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/pull-jobs/{job_id}": {
      "get": {
        "description": "Pull job status + progress (poll target for the GUI).",
        "operationId": "pull_job_status_api_missions_pull_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pull Job Status",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/pull-jobs/{job_id}/cancel": {
      "post": {
        "description": "Ask an in-flight pull to stop (GUI Cancel button / CLI Ctrl-C).\n\nSets the job's cooperative-cancel flag; the background worker checks it (per download event,\nthrottled) and aborts before install, recording `cancelled`. Returns the current job view so\nthe caller sees `cancel_requested` immediately. Idempotent: cancelling an already-terminal job\n(it finished racing the cancel) returns its view unchanged. 404 if the job is unknown.",
        "operationId": "cancel_pull_job_api_missions_pull_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Pull Job",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}": {
      "delete": {
        "description": "Uninstall an installed mission \u2014 local (your_own) or pulled (library).\n\nRemoves the local install directory so the mission leaves Your Own / drops back to\nnot-installed in the library. 404 if nothing is installed under that id. Routes the write\nthrough the server's install root (single writer, parity with ingest/pull). The fused image is\nintentionally left in the local store; pruning it is a separate concern.",
        "operationId": "delete_mission_api_missions__mission_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Mission",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/manifest": {
      "get": {
        "description": "The full mission.json for an id \u2014 Your Own copy if installed, else the library.",
        "operationId": "mission_manifest_api_missions__mission_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mission Manifest",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/pull": {
      "post": {
        "description": "Pull a free-library mission (image-pull) so it becomes selectable/runnable.",
        "operationId": "pull_api_missions__mission_id__pull_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pull",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/pull-jobs": {
      "post": {
        "description": "Start (or join) a background pull for a library mission (202 + job_id, idempotent).\n\nThe image download can take minutes \u2014 far past what a browser will hold one request open\nfor \u2014 so it runs on a background task; poll GET /missions/pull-jobs/{job_id}. A POST\nwhile a pull for this mission is active returns the SAME job_id without starting a\nsecond worker (server-side double-pull guard).",
        "operationId": "start_pull_job_api_missions__mission_id__pull_jobs_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullJobStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Pull Job",
        "tags": [
          "missions"
        ]
      }
    },
    "/api/missions/{mission_id}/terrain": {
      "get": {
        "description": "The mission's projected v2 terrain \u2014 the same base graph a run of it starts from.\n\nRuns the run-scoped projector over the manifest (installed copy, else the library) with\nno run attached: infra scaffold + authored/static-ips mission plane, every element in its\nbase state, no updates. This is what the mission detail page draws so the ACTUAL terrain\nmap \u2014 not a simplification \u2014 is visible before a run (or a pull). Unknown id \u2192 404.",
        "operationId": "mission_terrain_api_missions__mission_id__terrain_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedTerrainV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mission Terrain",
        "tags": [
          "missions"
        ]
      }
    },
...
  }
}